Moved CoreDataFeedStore and DataModel to production
